finished level transition

+++ src/actor.c
@@ -132,10 +132,11 @@ void actor_draw(Actor *actor)
     {
         return;
     }
+
     float sprite_x = actor->position.x + actor->sprite_offset.x;
     float sprite_y = actor->position.y + actor->sprite_offset.y;
     int flags = actor->is_facing_right ? 0 : ALLEGRO_FLIP_HORIZONTAL;
-    al_draw_bitmap(actor->sprite, (int)sprite_x, (int)sprite_y, flags);
+    al_draw_bitmap(actor->sprite, sprite_x, sprite_y, flags);
 
     // al_draw_rectangle(actor->position.x, actor->position.y, actor->position.x + actor->width, actor->position.y + actor->height, al_map_rgb(255, 0, 0), 1.0f);
 }
